fix: refactor routes and remove DetailPage component
