+++ web/src/App.tsx
@@ -2,7 +2,6 @@ import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import { FC } from 'react';
 import MainPage from './components/pages/MainPage';
 import PresentationPage from './components/pages/PresentationPage';
-import DetailPage from './components/pages/DetailPage';
 
 const App: FC = () => {
   return (
@@ -11,7 +10,7 @@ const App: FC = () => {
         <Routes>
           <Route path='/' element={<MainPage />} />
           <Route path='/presentation' element={<PresentationPage />} />
-          <Route path='/slides/:uuid' element={<DetailPage />} />
+          <Route path='/:uuid' element={<MainPage />} />
           <Route path='*' element={<Navigate to='/' />} />
         </Routes>
       </BrowserRouter>
