+++ web/src/components/pages/MainPage.tsx
@@ -1,31 +1,40 @@
-import { FC } from 'react';
-import { useSelector } from 'react-redux';
+import { FC, useEffect } from 'react';
+import useSWR from 'swr';
+import { useDispatch, useSelector } from 'react-redux';
 
-import Layout from '../templates/Layout';
 import { RootState } from '../../store';
+import { fetchInstanceWithToken } from '../../libs/fetchInstance';
+import { contentSlice } from '../../features/content';
+import { useParams, useNavigate } from 'react-router-dom';
 
-import ModeSwitchBar from '../organisms/ModeSwitchBar';
-import ActionPanel from '../organisms/ActionPanel';
-import EditArea from '../organisms/EditArea';
-import ViewArea from '../organisms/ViewArea';
+import EditorAndViewer from '../organisms/EditorAndViewer';
 
 const MainPage: FC = () => {
+  const navigate = useNavigate();
+  const dispatch = useDispatch();
+  const { uuid } = useParams<{ uuid: string }>();
   const { mode } = useSelector((state: RootState) => state.view);
+  const { token } = useSelector((state: RootState) => state.user);
 
-  return (
-    <Layout>
-      <div className='h-full container mx-auto px-4 py-4'>
-        <div className='flex justify-between items-start flex-col gap-2 md:flex-row md:items-end'>
-          <ModeSwitchBar />
-          <ActionPanel />
-        </div>
-        <div className='h-5/6 flex gap-4 py-4'>
-          {mode !== 'view' && <EditArea />}
-          {mode !== 'edit' && <ViewArea />}
-        </div>
-      </div>
-    </Layout>
-  );
+  if (!token) navigate('/');
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await fetchInstanceWithToken(token).get(`/slides/${uuid}`);
+        const data = response.data;
+
+        dispatch(contentSlice.actions.setContent(data.content));
+      } catch (error) {
+        navigate('/');
+      }
+    };
+
+    if (uuid) fetchData();
+  }, [dispatch, navigate, token, uuid]);
+
+  if (!uuid) return <EditorAndViewer mode={mode} />;
+  return <EditorAndViewer mode={mode} />;
 };
 
 export default MainPage;
